feat(engine): Convert the file loader to a better file system.
The current file loader was not providing enough customization possibilities during integration of Amplitude in games or game engines. Now with the new file system architecture, it is possible to implement a custom class that may browse for Amplitude assets from an archive or bundle file, or any other mean that suits the need of the user.