fix(allocator): Revert usage of custom heaps.
This was causing the allocator to throw errors when a program tried to allocate memory from a heap in a thread where the heap itself is not initialized. For now it's safer to let mi_malloc create and destroy the default heap foreach treads, as it is built for that.

Signed-off-by: Axel Nana <axel.nana@aliens-group.com>